requirements.txt
streamlit
langchain
google-generativeai
python-dotenv
PyPDF2
langchain-community
sentence-transformers
langchain-huggingface
faiss-cpu